lib: move lvgl fs glue to lvgl_addons lib
