+++ lib/lvgl_addons.scons
@@ -23,7 +23,7 @@ libenv.AppendUnique(
 sources = []
 sources += libenv.Glob(LIB_FOLDER + "/fonts/src/*.c", source=True)
 sources += libenv.Glob(LIB_FOLDER + "/themes/*.c", source=True)
-# sources += libenv.Glob(LIB_FOLDER + "fs/src/*.c", source=True)
+sources += libenv.Glob(LIB_FOLDER + "/fs/*.c", source=True)
 
 lib = libenv.StaticLibrary("${FW_LIB_NAME}", sources)
 libenv.Install("${LIB_DIST_DIR}", lib)
